[CI] Migrate CUDA Graph tests to test/registered/cuda_graph/
Part of #13808

This PR migrates piecewise CUDA graph tests to the new registry-based
test infrastructure, organized by GPU memory requirements:

**Changes:**
- Move tests from `test/srt/` to `test/registered/cuda_graph/`
- Split tests by GPU size:
  - `test_piecewise_cuda_graph_small_1_gpu.py` (24GB GPU) - 460s
  - `test_piecewise_cuda_graph_large_1_gpu.py` (80GB GPU) - 480s
  - `test_piecewise_cuda_graph_2_gpu.py` (2x 80GB GPU) - 200s
- Add `stage-b-test-large-1-gpu` suite and workflow job
- Rename `stage-b-test-2-gpu` to `stage-b-test-large-2-gpu` for consistency
- Fix `test_lora_tp.py` suite name (was using invalid suite)
- Update slash command handler for `/rerun-stage` support

**Test organization by GPU:**
- Small 1-GPU (24GB): Correctness, Benchmark, FP4, DeepSeek, FP8, VL models
- Large 1-GPU (80GB): Qwen3 30B MoE, GPTQ 30B, AWQ 32B
- 2-GPU (80GB): Qwen3 Omni MoE (tp=2)
