+++ test/registered/lora/test_lora_tp.py
@@ -28,7 +28,7 @@ from sglang.test.lora_utils import (
 )
 from sglang.test.test_utils import CustomTestCase, is_in_ci
 
-register_cuda_ci(est_time=116, suite="stage-b-test-small-2-gpu")
+register_cuda_ci(est_time=116, suite="stage-b-test-large-2-gpu")
 
 
 class TestLoRATP(CustomTestCase):
