--- test/srt/test_piecewise_cuda_graph_1_gpu_a.py
@@ -1,175 +1,0 @@
-import unittest
-
-from sglang.srt.utils import get_device_sm, kill_process_tree
-from sglang.test.few_shot_gsm8k import run_eval as run_eval_few_shot_gsm8k
-from sglang.test.run_eval import run_eval
-from sglang.test.test_utils import (
-    DEFAULT_MODEL_NAME_FOR_TEST,
-    DEFAULT_MODEL_NAME_FOR_TEST_MLA,
-    DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
-    DEFAULT_URL_FOR_TEST,
-    CustomTestCase,
-    SimpleNamespace,
-    popen_launch_server,
-    run_bench_one_batch,
-)
-
-
-class TestPiecewiseCudaGraphCorrectness(CustomTestCase):
-    @classmethod
-    def setUpClass(cls):
-        cls.model = DEFAULT_MODEL_NAME_FOR_TEST
-        cls.base_url = DEFAULT_URL_FOR_TEST
-        cls.process = popen_launch_server(
-            cls.model,
-            cls.base_url,
-            timeout=DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
-            other_args=["--enable-piecewise-cuda-graph"],
-        )
-
-    @classmethod
-    def tearDownClass(cls):
-        kill_process_tree(cls.process.pid)
-
-    def test_mmlu(self):
-        args = SimpleNamespace(
-            base_url=self.base_url,
-            model=self.model,
-            eval_name="mmlu",
-            num_examples=64,
-            num_threads=32,
-        )
-
-        metrics = run_eval(args)
-        self.assertGreaterEqual(metrics["score"], 0.65)
-
-
-class TestPiecewiseCudaGraphBenchmark(CustomTestCase):
-
-    def test_latency(self):
-        prefill_latency, _, _ = run_bench_one_batch(
-            DEFAULT_MODEL_NAME_FOR_TEST,
-            other_args=["--enable-piecewise-cuda-graph"],
-        )
-        self.assertLess(prefill_latency, 0.015)
-
-
-@unittest.skipIf(get_device_sm() < 100, "Test requires CUDA SM 100 or higher")
-class TestPiecewiseCudaGraphLlama31FP4(CustomTestCase):
-    """MGSM test: piecewise CUDA graph with NVFP4 Llama3.1 8B on Blackwell."""
-
-    @classmethod
-    def setUpClass(cls):
-        cls.model = "nvidia/Llama-3.1-8B-Instruct-FP4"
-        cls.base_url = DEFAULT_URL_FOR_TEST
-        cls.process = popen_launch_server(
-            cls.model,
-            cls.base_url,
-            timeout=DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
-            other_args=[
-                "--enable-piecewise-cuda-graph",
-                "--quantization",
-                "modelopt_fp4",
-                "--mem-fraction-static",
-                "0.8",
-            ],
-        )
-
-    @classmethod
-    def tearDownClass(cls):
-        kill_process_tree(cls.process.pid)
-
-    def test_mgsm_accuracy(self):
-        num_examples = 1319
-        args = SimpleNamespace(
-            base_url=self.base_url,
-            model=self.model,
-            eval_name="mgsm_en",
-            num_examples=num_examples,
-            num_threads=min(num_examples, 1024),
-        )
-        metrics = run_eval(args)
-        print(f"MGSM Accuracy: {metrics['score']:.3f}")
-        self.assertGreaterEqual(metrics["score"], 0.78)
-
-
-class TestPiecewiseCudaGraphQwen3MoE(CustomTestCase):
-    """Test piecewise CUDA graph with Qwen3-Coder-30B-A3B-Instruct MoE model"""
-
-    @classmethod
-    def setUpClass(cls):
-        cls.model = "Qwen/Qwen3-Coder-30B-A3B-Instruct"
-        cls.base_url = DEFAULT_URL_FOR_TEST
-        cls.process = popen_launch_server(
-            cls.model,
-            cls.base_url,
-            timeout=DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
-            other_args=[
-                "--enable-piecewise-cuda-graph",
-                "--piecewise-cuda-graph-compiler",
-                "eager",
-            ],
-        )
-
-    @classmethod
-    def tearDownClass(cls):
-        kill_process_tree(cls.process.pid)
-
-    def test_gsm8k_accuracy(self):
-        """Test GSM8K accuracy with 8-shot setting"""
-        num_examples = 2000
-
-        args = SimpleNamespace(
-            base_url=self.base_url,
-            model=self.model,
-            eval_name="mgsm_en",
-            num_examples=num_examples,
-            num_threads=min(num_examples, 1024),
-        )
-
-        metrics = run_eval(args)
-        print(f"GSM8K Accuracy: {metrics['score']:.3f}")
-
-        self.assertGreaterEqual(metrics["score"], 0.90)
-
-
-class TestPiecewiseCudaGraphDeepSeek(CustomTestCase):
-    @classmethod
-    def setUpClass(cls):
-        cls.model = DEFAULT_MODEL_NAME_FOR_TEST_MLA
-        cls.base_url = DEFAULT_URL_FOR_TEST
-        cls.process = popen_launch_server(
-            cls.model,
-            cls.base_url,
-            timeout=DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
-            other_args=[
-                "--enable-piecewise-cuda-graph",
-                "--piecewise-cuda-graph-compiler",
-                "eager",
-                "--piecewise-cuda-graph-max-tokens",
-                "4096",  # should less than max_context_len
-            ],
-        )
-
-    @classmethod
-    def tearDownClass(cls):
-        kill_process_tree(cls.process.pid)
-
-    def test_gsm8k(self):
-        args = SimpleNamespace(
-            num_shots=5,
-            data_path=None,
-            num_questions=200,
-            max_new_tokens=512,
-            parallel=128,
-            host="http://127.0.0.1",
-            port=int(self.base_url.split(":")[-1]),
-        )
-        metrics = run_eval_few_shot_gsm8k(args)
-        print(metrics)
-
-        self.assertGreater(metrics["accuracy"], 0.62)
-
-
-if __name__ == "__main__":
-    unittest.main()
